Add files via upload
This was all done on a Raspberry Pi 3.  (the 4 apparently have problems running the files and will update if i find out more)
Open the README file and follow the steps to install the programs and files that are needed. After you install and run all the steps, you should have ran ./run.sh once that runs it should display Starting KerberosSDR along with a python server running and PHP server. Enter the following into your browser and you should be given a selection for the compass: localhost:8081 .
If you are given the error usb_claim_interface error -6, follow this link for help 
If you are given the error usb_open error -3, this was the link I found for help: https://github.com/merbanan/rtl_433/issues/740
This link is the RTL_SDR quick start guide for reference: https://www.rtl-sdr.com/ksdr/
This link was the original link for the kerberos: https://github.com/rtlsdrblog/kerberossdr/